client_ssl::ssl_read_cb: fix warnings

+++ src/ssl.cpp
@@ -35,7 +35,7 @@ static int ssl_bio_read(BIO* bio, char* data, int len) noexcept {
 	auto& c = *(ws::client_ssl*)BIO_get_data(bio);
 
 	try {
-		return c.ssl_read_cb(span((uint8_t*)data, len));
+		return (int)c.ssl_read_cb(span((uint8_t*)data, len));
 	} catch (...) {
 		c.exception = current_exception();
 		return -1;
@@ -287,8 +287,8 @@ static void add_certs_to_store(X509_STORE* store) {
 #endif
 
 namespace ws {
-	int client_ssl::ssl_read_cb(std::span<uint8_t> s) {
-		int copied = 0;
+	size_t client_ssl::ssl_read_cb(std::span<uint8_t> s) {
+		size_t copied = 0;
 
 		if (s.empty())
 			return 0;
@@ -306,7 +306,7 @@ namespace ws {
 			s = s.subspan(to_copy);
 		}
 
-		auto ret = ::recv(client.sock, (char*)s.data(), s.size(), 0);
+		auto ret = ::recv(client.sock, (char*)s.data(), (int)s.size(), 0);
 
 #ifdef _WIN32
 		if (ret == SOCKET_ERROR) {
